Split references.rs to references.rs and annotations.rs
