+++ crates/sierra_to_casm/src/lib.rs
@@ -1,5 +1,6 @@
 //! CASM backend. Compiles from Sierra down to CASM. See [sierra] and [casm]
 
+pub mod annotations;
 pub mod compiler;
 pub mod invocations;
 pub mod references;
